Renamged graphics.py to game_renderer.py
