+++ game_renderer.py
